new problems (but unsolved)
my reflection:
Why did I not get the problems right?
It is because I did not use any of the algorithms I learned. I am just trying to solve all problems with simple loops.
I need to implement the search algorithms I learned into the code, so that the program runs with shorter time and more efficient memory.